mempool: add unique mempool option
A unique mempool uses a Map internally to make sure we do
not add pointers twice.  This is used by portfmt.  We do this in
Mempool and not in stack_contains() or queue_contains() with a Set
to not make their performance worse.  We would need to call both
set_add() and set_remove() there to maintain the set which would
have a significant impact on both time and memory.