map: Use red-black tree instead of splay tree
Even simple lookups cause memory writes to happen with splay trees.
This can cause problems when you try to read a Map from multiple
threads as we do in portscan.  With red-black trees this should be
fine as the tree is only modified when you insert or remove an item.
It does not really seem to impact performance of portscan either.